add healthcheck in testsuite for armada server (#1120)
* add healthcheck in testsuite for armada server

* move rest healthcheck to client module

* add armada rest server healthcheck to testsuite cmd; lint testsuite cmd module

* run goimports on cmd/testusite module

* minor lint; move armada healthcheck from initParams to testCmdRunE function

* fix goimports lint issues in cmd/testsuite/cmd module

* capitalize Armada in code; return tuple for healthcheck call

* capitalize Armada in code; return tuple for healthcheck call

* wrap error in WithStack
