+++ pkg/client/connection.go
@@ -1,8 +1,13 @@
 package client
 
 import (
+	"fmt"
+	"net/http"
 	"strings"
 	"time"
+
+	"github.com/pkg/errors"
+	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/G-Research/armada/pkg/client/auth/exec"
 
@@ -17,6 +22,7 @@ import (
 
 type ApiConnectionDetails struct {
 	ArmadaUrl                   string
+	ArmadaRestUrl               string
 	BasicAuth                   common.LoginCredentials
 	OpenIdAuth                  oidc.PKCEDetails
 	OpenIdDeviceAuth            oidc.DeviceDetails
@@ -97,5 +103,28 @@ func transportCredentials(config *ApiConnectionDetails) grpc.DialOption {
 	if !config.ForceNoTls && !strings.Contains(config.ArmadaUrl, "localhost") {
 		return grpc.WithTransportCredentials(credentials.NewClientTLSFromCert(nil, ""))
 	}
-	return grpc.WithInsecure()
+	return grpc.WithTransportCredentials(insecure.NewCredentials())
+}
+
+// ArmadaHealthCheck calls Armada Server /health endpoint.
+//
+// Returns true if response status code is in range [200-399], otherwise returns false.
+func (a *ApiConnectionDetails) ArmadaHealthCheck() (ok bool, err error) {
+	url := a.ArmadaRestUrl
+	if url == "" {
+		return false, errors.New("Armada server rest api url not provided")
+	}
+	if !strings.HasPrefix(url, "http") {
+		url = fmt.Sprintf("http://%s", url)
+	}
+	healthEndpoint := fmt.Sprintf("%s/health", url)
+	resp, err := http.Get(healthEndpoint)
+	if err != nil {
+		return false, errors.WithStack(err)
+	}
+	if resp.StatusCode < 200 || resp.StatusCode > 399 {
+		return false, nil
+	}
+
+	return true, nil
 }
